+++ src/main/java/solutions/fairdata/openrefine/metadata/dto/config/ProjectConfigDTO.java
@@ -20,28 +20,20 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package solutions.fairdata.openrefine.metadata.commands.response.config;
+package solutions.fairdata.openrefine.metadata.dto.config;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import solutions.fairdata.openrefine.metadata.dto.config.ProjectConfigDTO;
-import solutions.fairdata.openrefine.metadata.dto.config.SettingsConfigDTO;
 
-@NoArgsConstructor
+import java.util.HashMap;
+
 @AllArgsConstructor
+@NoArgsConstructor
 @Getter
 @Setter
-public class SettingsResponse {
-
-    private String status;
-    private SettingsConfigDTO settings;
-    private ProjectConfigDTO projectData;
-
-    public SettingsResponse(SettingsConfigDTO settings, ProjectConfigDTO projectData) {
-        this.status = "ok";
-        this.settings = settings;
-        this.projectData = projectData;
-    }
+public class ProjectConfigDTO {
+    private HashMap<String, String> lastCatalog = new HashMap<>();
+    private HashMap<String, String> lastDataset = new HashMap<>();
 }