+++ src/main/java/solutions/fairdata/openrefine/metadata/model/MetadataOverlayModel.java
@@ -20,28 +20,36 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package solutions.fairdata.openrefine.metadata.commands.response.config;
+package solutions.fairdata.openrefine.metadata.model;
 
+import com.google.refine.model.OverlayModel;
+import com.google.refine.model.Project;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import solutions.fairdata.openrefine.metadata.dto.config.ProjectConfigDTO;
-import solutions.fairdata.openrefine.metadata.dto.config.SettingsConfigDTO;
 
 @NoArgsConstructor
 @AllArgsConstructor
 @Getter
 @Setter
-public class SettingsResponse {
+public class MetadataOverlayModel implements OverlayModel {
 
-    private String status;
-    private SettingsConfigDTO settings;
     private ProjectConfigDTO projectData;
 
-    public SettingsResponse(SettingsConfigDTO settings, ProjectConfigDTO projectData) {
-        this.status = "ok";
-        this.settings = settings;
-        this.projectData = projectData;
+    @Override
+    public void onBeforeSave(Project project) {
+        // noop
+    }
+
+    @Override
+    public void onAfterSave(Project project) {
+        // noop
+    }
+
+    @Override
+    public void dispose(Project project) {
+        // noop
     }
 }